requirements.txt
absl-py==0.13.0
alabaster==0.7.12
alembic==1.4.1
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
blis==0.7.4
Boruta==0.3
Bottleneck==1.3.2
branca==0.4.2
Brotli==1.0.9
cached-property==1.5.2
cachetools==4.2.2
catalogue==1.0.0
catboost==0.26
certifi==2020.12.5
cffi==1.14.5
cfgv==3.3.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorlover==0.3.0
cufflinks==0.17.3
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
dash==1.20.0
dash-bootstrap-components==0.12.2
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
databricks-cli==0.14.3
dataframe==0.2.1.3
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
dnspython==1.16.0
docker==5.0.0
docutils==0.17.1
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
Flask==2.0.0
Flask-Compress==1.9.0
flatbuffers==1.12
fluidsynth==0.2
folium==0.12.1
ftfy==6.0.3
funcy==1.16
future==0.18.2
gast==0.4.0
gensim==3.8.3
geocoder==1.38.1
geographiclib==1.50
geopy==2.1.0
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googletrans==3.0.0
graphviz==0.16
greenlet==1.1.0
grpcio==1.34.1
h11==0.9.0
h2==3.2.0
h2o==3.32.1.3
h5py==3.1.0
hpack==3.0.0
hstspreload==2020.12.22
htmlmin==0.1.12
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==2.2.6
idna==2.10
ImageHash==4.2.0
imageio==2.9.0
imagesize==1.2.0
imbalanced-learn==0.7.0
importlib-metadata==4.0.1
iniconfig==1.1.1
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.0
joblib==1.0.1
jsons==1.4.2
jsonschema==3.2.0
jupyter-client==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kmodes==0.11.0
lightgbm==3.2.1
llvmlite==0.36.0
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.4.1
midi2audio==0.1.1
MIDIUtil==1.2.1
mido==1.2.10
missingno==0.4.2
mistune==0.8.4
mlflow==1.18.0
mlxtend==0.18.0
more-itertools==8.8.0
multimethod==1.4
murmurhash==1.0.5
music21==6.7.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
nodeenv==1.6.0
nose==1.3.7
notebook==6.1.5
numba==0.53.1
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.54
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandas-profiling==3.0.0
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
phik==0.11.2
pickleshare==0.7.5
Pillow==8.2.0
plac==1.1.3
plotly==4.14.3
plotly-express==0.4.1
pluggy==0.13.1
pre-commit==2.13.0
preshed==3.0.5
pretty-html-table==0.9.5
pretty-midi==0.2.9
prometheus-client==0.10.1
prometheus-flask-exporter==0.18.2
prompt-toolkit==3.0.18
protobuf==3.17.0
py==1.10.0
py-translate==1.0.3
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycaret==2.3.1
pycparser==2.20
pydantic==1.8.2
pydeck==0.6.2
pygame==2.0.1
Pygments==2.8.1
pyLDAvis==3.3.1
pymongo==3.11.4
PyMySQL==1.0.2
pynndescent==0.5.2
pyod==0.8.9
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
querystring-parser==1.2.4
ratelim==0.1.6
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.5.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smart-open==5.1.0
smmap==4.0.0
sniffio==1.2.0
snowballstemmer==2.1.0
soupsieve==2.2.1
spacy==2.3.7
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.15
sqlparse==0.4.1
squarify==0.4.3
srsly==1.0.5
statsmodels==0.12.2
streamlit==0.82.0
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.5
testpath==0.5.0
textblob==0.15.3
thinc==7.4.5
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
toolz==0.11.1
tornado==6.1
tox==3.23.1
tqdm==4.60.0
traitlets==5.0.5
translate==3.5.0
tweepy==3.10.0
typing-extensions==3.7.4.3
typish==1.9.2
tzlocal==2.1
umap-learn==0.5.1
urllib3==1.26.4
validators==0.18.2
vega-datasets==0.9.0
virtualenv==20.4.7
visions==0.7.1
waitress==2.0.0
wasabi==0.8.2
watchdog==2.1.1
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.12.1
xgboost==1.4.2
xlrd==2.0.1
yellowbrick==1.3.post1
zipp==3.4.1
